+++ packages/saved-views-client/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/iTwin/saved-views/tree/HEAD/packages/saved-views-client)
+
+### Fixes
+
+* Fix `ITwinSavedViewsClient.updateSavedView` failing when saved view data contains URL fields
+
 ## [0.2.0](https://github.com/iTwin/saved-views/tree/v0.2.0-client/packages/saved-views-client) - 2024-04-08
 
 ### Breaking changes